requirements.txt
django>=4.2
psycopg2-binary
djangorestframework
gunicorn
celery
redis
django-celery-results
